refactored ui code to abstract qt designer generated code wrapped qt designer classes and integrated extended ui functionality added focus for primary text fields in dialog boxes smoked a cuban cigar =)
