+++ .vscode/settings.json
@@ -5,5 +5,10 @@
     ],
     "workbench.colorCustomizations": {
         "editorRuler.foreground": "#ffb24d80"
-    }
+    },
+    "python.testing.pytestArgs": [
+        "test"
+    ],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true
 }
