--- ui/__init__.py
@@ -1,3 +1,0 @@
-from .ui_form_main import Ui_MainWindow
-from .ui_dialog_login import Ui_DialogLogin
-from .ui_dialog_expense import Ui_DialogExpense
